requirements.txt
streamlit
yfinance
matplotlib
pandas
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-prometheus
prometheus-client